(#2065) Modal Unit Tests
Closes #2065
