+++ packages/ncids-js/src/components/usa-modal/__tests__/modal.create.test.ts
@@ -167,50 +167,4 @@ describe('Modal - Create Test', () => {
 			expect(modal).toBeInstanceOf(USAModal);
 		});
 	});
-
-	it('Test body Class and padding', async () => {
-		Object.defineProperty(document.body.style, 'backgroundColor', {
-			writable: true,
-			value: '',
-		});
-
-		const setPropertySpy = jest.spyOn(
-			CSSStyleDeclaration.prototype,
-			'setProperty'
-		);
-		const removePropertySpy = jest.spyOn(
-			CSSStyleDeclaration.prototype,
-			'removeProperty'
-		);
-		setPropertySpy.mockImplementation(() => {
-			return '31px';
-		});
-		removePropertySpy.mockImplementation(() => {
-			return '';
-		});
-		const user = userEvent.setup();
-		const container = document.createElement('div');
-		container.innerHTML = modalDom;
-		document.body.append(container);
-
-		const element = document.querySelector('.usa-modal');
-		USAModal.create(<HTMLElement>element);
-
-		const openers = document.querySelectorAll('[data-open-modal]');
-
-		// Initial state: no padding and no modal class
-		expect(document.body).not.toHaveClass('usa-js-modal--active');
-		expect(document.body).not.toHaveStyle({});
-		await user.click(openers[0]);
-
-		// Check if the class for active set
-		expect(document.body).toHaveClass('usa-js-modal--active');
-
-		const closers = document.querySelectorAll('[data-close-modal]');
-		await user.click(closers[0]);
-
-		// Check if the padding and class are removed
-		expect(document.body).not.toHaveStyle({});
-		expect(document.body).not.toHaveClass('usa-js-modal--active');
-	});
 });
